+++ CHANGELOG.md
@@ -3,3 +3,12 @@
 
 ### v1.0.1
 * Updated installation instructions to use the mod manager instead.
+
+### v1.0.2
+* [Enforcer] Shrapnel Grenade
+  * Damage: 3x70% -> 3x75%
+* [Mercenary] Whirlwind
+  * Damage: 2x125% -> 2x130%
+* [Loader] Bullet Punch
+  * Damage: 1000% -> 1500%
+* Fixed incorrect skill descriptions for various characters.